Import image files and show in post card

+++ app/Http/Controllers/PostController.php
@@ -5,9 +5,16 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['auth'])->except(['index', 'show']);
+    }
+
     // Display list posts
     public function index()
     {
@@ -26,13 +33,25 @@ class PostController extends Controller
     public function store(Request $request, Post $post)
     {
         // Validate:
-        $fields = $request->validate([
+        $request->validate([
             'title' => ["required", "max:255"],
             'body' => ["required"],
+            'image' => ['nullable', 'file', 'max:3000', 'mimes:png,jpg,webp'],
         ]);
 
+        // Storage image if exists:
+        // Storage in [store/app/public/posts_images]
+        $path = null;
+        if ($request->hasFile('image')) {
+            $path = Storage::disk("public")->put("posts_images", $request->image);
+        }
+
         // Create a post
-        Auth::user()->posts()->create($fields);
+        Auth::user()->posts()->create([
+            'title' => $request->title,
+            'body' => $request->body,
+            'image' => $path,
+        ]);
 
         return back()->with("success", "Your post was created!");
     }
@@ -44,11 +63,17 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        // Authorize the action:
+        Gate::authorize("modify", $post);
+
         return view("posts.edit", ["post" => $post]);
     }
 
     public function update(Request $request, Post $post)
     {
+        // Authorize the action:
+        Gate::authorize("modify", $post);
+
         // Validate:
         $fields = $request->validate([
             'title' => ["required", "max:255"],
@@ -63,11 +88,12 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        // Authorize the action:
+        Gate::authorize("modify", $post);
+
         $post->delete();
 
         // Redirect to dashboard:
         return back()->with('delete', 'Your post was deleted!');
     }
-
-
 }
